+++ src/Vankosoft/ApplicationBundle/Controller/VankosoftIssueBoardController.php
@@ -1,0 +1,149 @@
+<?php namespace Vankosoft\ApplicationBundle\Controller;
+
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Form\FormInterface;
+use Sylius\Resource\Doctrine\Persistence\RepositoryInterface;
+
+use Vankosoft\ApplicationBundle\Component\Exception\VankosoftApiException;
+use Vankosoft\ApplicationBundle\Component\Application\ProjectIssue;
+use Vankosoft\ApplicationBundle\Form\ProjectIssueForm;
+
+class VankosoftIssueBoardController extends AbstractController
+{
+    /** @var ProjectIssue */
+    private $vsProject;
+    
+    public function __construct(
+        ProjectIssue $vsProject
+    ) {
+        $this->vsProject    = $vsProject;
+    }
+    
+    public function showKanbanboardAction( Request $request ): Response
+    {
+        $apiEnabled = $this->getParameter( 'vs_application.vankosoft_api.enabled' );
+        $apiBoard   = $this->getParameter( 'vs_application.vankosoft_api.kanbanboard' );
+        
+        if( ! $apiEnabled ) {
+            throw new VankosoftApiException( 'VankoSoft API is NOT Enabled !!! Please Enable it and Configure it !!!' );
+        }
+        
+        if ( $apiBoard === ProjectIssue::BOARD_UNDEFINED ) {
+            throw new VankosoftApiException( 'VankoSoft API Kanbanboard Slug is NOT Defined !!!' );
+        }
+        
+        $board = $this->vsProject->getKanbanboard( $apiBoard );
+        //echo '<pre>'; var_dump( $issues ); die;
+        
+        return $this->render( '@VSApplication/Pages/ProjectIssuesBoard/kanbanboard.html.twig', [
+            'board' => $board
+        ]);
+    }
+    
+    public function showTaskAction( $taskId, Request $request ): Response
+    {
+        $apiEnabled = $this->getParameter( 'vs_application.vankosoft_api.enabled' );
+        $apiBoard   = $this->getParameter( 'vs_application.vankosoft_api.kanbanboard' );
+        
+        if( ! $apiEnabled ) {
+            throw new VankosoftApiException( 'VankoSoft API is NOT Enabled !!! Please Enable it and Configure it !!!' );
+        }
+        
+        if ( $apiBoard === ProjectIssue::BOARD_UNDEFINED ) {
+            throw new VankosoftApiException( 'VankoSoft API Kanbanboard Slug is NOT Defined !!!' );
+        }
+        
+        $board = $this->vsProject->getKanbanboard( $apiBoard );
+        //echo '<pre>'; var_dump( $issues ); die;
+        
+        return $this->render( '@VSApplication/Pages/ProjectIssuesBoard/task.html.twig', [
+            'board' => $board
+        ]);
+    }
+    
+    public function moveTaskAction( $taskId, $pipelineId, Request $request ): Response
+    {
+        $apiEnabled = $this->getParameter( 'vs_application.vankosoft_api.enabled' );
+        $apiBoard   = $this->getParameter( 'vs_application.vankosoft_api.kanbanboard' );
+        
+        if( ! $apiEnabled ) {
+            throw new VankosoftApiException( 'VankoSoft API is NOT Enabled !!! Please Enable it and Configure it !!!' );
+        }
+        
+        if ( $apiBoard === ProjectIssue::BOARD_UNDEFINED ) {
+            throw new VankosoftApiException( 'VankoSoft API Kanbanboard Slug is NOT Defined !!!' );
+        }
+        
+        $board = $this->vsProject->getKanbanboard( $apiBoard );
+        //echo '<pre>'; var_dump( $issues ); die;
+        
+        return $this->render( '@VSApplication/Pages/ProjectIssuesBoard/task.html.twig', [
+            'board' => $board
+        ]);
+    }
+    
+    public function assignMemberAction( $taskId, $memberId, Request $request ): Response
+    {
+        $apiEnabled = $this->getParameter( 'vs_application.vankosoft_api.enabled' );
+        $apiBoard   = $this->getParameter( 'vs_application.vankosoft_api.kanbanboard' );
+        
+        if( ! $apiEnabled ) {
+            throw new VankosoftApiException( 'VankoSoft API is NOT Enabled !!! Please Enable it and Configure it !!!' );
+        }
+        
+        if ( $apiBoard === ProjectIssue::BOARD_UNDEFINED ) {
+            throw new VankosoftApiException( 'VankoSoft API Kanbanboard Slug is NOT Defined !!!' );
+        }
+        
+        $board = $this->vsProject->getKanbanboard( $apiBoard );
+        //echo '<pre>'; var_dump( $issues ); die;
+        
+        return $this->render( '@VSApplication/Pages/ProjectIssuesBoard/task.html.twig', [
+            'board' => $board
+        ]);
+    }
+    
+    public function createTaskAction( $pipelineId, Request $request ): Response
+    {
+        $apiEnabled = $this->getParameter( 'vs_application.vankosoft_api.enabled' );
+        $apiBoard   = $this->getParameter( 'vs_application.vankosoft_api.kanbanboard' );
+        
+        if( ! $apiEnabled ) {
+            throw new VankosoftApiException( 'VankoSoft API is NOT Enabled !!! Please Enable it and Configure it !!!' );
+        }
+        
+        if ( $apiBoard === ProjectIssue::BOARD_UNDEFINED ) {
+            throw new VankosoftApiException( 'VankoSoft API Kanbanboard Slug is NOT Defined !!!' );
+        }
+        
+        $board = $this->vsProject->getKanbanboard( $apiBoard );
+        //echo '<pre>'; var_dump( $issues ); die;
+        
+        return $this->render( '@VSApplication/Pages/ProjectIssuesBoard/task.html.twig', [
+            'board' => $board
+        ]);
+    }
+    
+    public function getAttachmentFormAction( $taskId, Request $request ): Response
+    {
+        $apiEnabled = $this->getParameter( 'vs_application.vankosoft_api.enabled' );
+        $apiBoard   = $this->getParameter( 'vs_application.vankosoft_api.kanbanboard' );
+        
+        if( ! $apiEnabled ) {
+            throw new VankosoftApiException( 'VankoSoft API is NOT Enabled !!! Please Enable it and Configure it !!!' );
+        }
+        
+        if ( $apiBoard === ProjectIssue::BOARD_UNDEFINED ) {
+            throw new VankosoftApiException( 'VankoSoft API Kanbanboard Slug is NOT Defined !!!' );
+        }
+        
+        $board = $this->vsProject->getKanbanboard( $apiBoard );
+        //echo '<pre>'; var_dump( $issues ); die;
+        
+        return $this->render( '@VSApplication/Pages/ProjectIssuesBoard/task.html.twig', [
+            'board' => $board
+        ]);
+    }
+}